// src/app/core/services/dto/character-dto.ts
/**
 * DTO model for character
 */
export interface CharacterDTO {
  /** Eye color */
  eye_color: string;

  /** Homeworld id */
  homeworld: number;

  /** Gender */
  gender: string;

  /** Height */
  height: string;

  /** Mass */
  mass: string;

  /** Name */
  name: string;

  /** Skin color */
  skin_color: string;

  /** Birth year */
  birth_year: string;
}